Add thoroughly documented suggested reconcile results.
The intention behind these variables is to make the intent of each reconcile
function's return value more obvious. It's difficult to determine elegant names
for these variables, so we lean heavily on their godoc strings to explain their
purposes and call out edge cases that will likely not be immediately obvious to
Crossplane contributors.

Signed-off-by: Nic Cope <negz@rk0n.org>