project setup: add submodule management to Makefile
Before crossplane can be built locally, the git submodules must be initialized.
Later, a developer may want to update the submodules again by hand. Originally,
this was explained in the documentation, and the command could be copied and
pasted into the terminal. It's more convenient to have a recipe in the Makefile
to do this instead.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>