make it possible to debug the core process of the extension manager
The extension manager detects the name of the running pod and namespace
so that it can fetch this kubernetes object and pluck the container
image out.  The container image is then reused in unpack jobs.

This PR bypasses the image lookup in the presence of an environment
variable.  This allows the parent process to be debugged because the
normal POD identification fails when run in a debugger.  There is
no pod when debugging, there is just a process connecting to the
Kubernetes API.

This PR does not assist in debugging the unpacker command of the
extension manager.

In VSCode, the Debug Configuration looks like this:

```json
        {
            "name": "Launch Extension Manager in Default K8s Context",
            "type": "go",
            "request": "launch",
            "mode": "auto",
            "program": "${workspaceFolder}/cmd/crossplane/main.go",
            "env": {
                "KUBECONFIG": "${env:HOME}/.kube/config",
                "POD_NAME": "crossplane-debug",
                "POD_NAMESPACE": "crossplane-system",
                "UNPACK_IMAGE": "crossplane/crossplane"
            },
            "args": ["extension", "manage", "--debug", "--sync=60s"]
        }
```

Signed-off-by: Marques Johansson <marques@upbound.io>