aws provider setup: use explicit `--output json`
In some of the aws provider setup script, the commands use `jq` to parse
json, but without explicitly telling the `aws` command to format its
output as json. If a user has a different default (such as `text`, which
is common), the script would previously print non-obvious errors. Now,
it will Just Work, because it explicitly asks for json and overrides the
configured default.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>