wordpress gcp example: whitelist all networks for cloudsql access
In order for an application to access a cloudsql instance, it must be
able to connect to it. Previously, this was done by hard-coding a rule
to allow all network connections to the instance. That behavior was
removed in favor of the `authorizedNetworks` field, but the wordpress
example was not using it, so there was no network connectivity. This
change uses the allow-all rule again, but in the wordpress example's
yaml files this time. This change is the simplest connectivity fix so
that the wordpress gcp example just works out of the box.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>