design: add one pager on developer experience mindset
The goal of adding a document about developer experience mindset is to
explain and document our thinking around developer experiences. This
allows us to more easily keep the mindset in mind, teach new
contributors what our mindset is, and share our mindset with people who
may be interested in it.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>