cli design: clarify how versioning and install work
Since the document was originally written, the idea of what support for
an `install` command may look like has changed, so the document needs to
be updated. There was also some ambiguity about how versioning would be
handled, so that's been clarified. This commit also reformats long lines
so that they're not long anymore.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>