stacks guide: add links
We want to have a lot of links to other documents for users to read if
they want to learn more about something. This change adds a bunch of
those links to the content. Not all of them have destinations yet, but
they should all or mostly be added now at least.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>