stacks: fixes for unpacking stacks
This commit ensures that the unpacking process starts with a base
dir that includes the .registry folder.  This matters now since
we look for top level icons only in the base dir, so knowing the
right base dir is important.

The troubleshooting ability of the stack manager is also increased
by adding the path to error messages when we fail to parse a
resource.

Finally, we also use filepath.Clean() to ensure that the path being
passed in as the base dir is always clean and normalized.

Signed-off-by: Jared Watts <jbw976@gmail.com>