initial one pager thinking for stack templating
This PR introduces the concept and benefits of Template Stacks as well
as exploring a proposed Template Stack Manager which may facilitate
this experience. Implementation challenges for a Template Stack Manager
and approaches to overcome them are provided.

Throughout the design, Go templates are used to illustrate potential
usage, but this is an implementation detail that will be the subject of
additional design documents.

Addresses #877

Signed-off-by: Marques Johansson <marques@upbound.io>