Remove low value API tests
This commit removes two classes of test that I consider to be of low value:

Schema registration tests. This class of test tests that the AddToScheme
functions that register API types with the Kubernetes scheme is working. I
consider this test to be low value for two reasons. Firstly, it's testing a very
small, generated scaffold around runtime.SchemeBuilder. Crossplane should not be
testing the logic of the runtime package; that's the runtime package's job.
Secondly, the intent of the tests seems to be to ensure that we remember to
register all of the API types that we intended to register; i.e. that we do not
add a new API type and forget to register it with the scheme. Unfortunately for
the test to catch this case we'd have to remember to update the test itself, but
not the AddToSchemes slice.

"Storage" tests. These appear to be variants on old tests scaffolded by
kubebuilder. Each test starts up a Kubernetes API server then ensures it can
CRUD the API type under test against that API server. As far as I can tell this
is just testing that our API types are in fact valid Kubernetes types, i.e. that
they have type and object metadata.

Signed-off-by: Nic Cope <negz@rk0n.org>