docs: add definition of done
In the retrospective for the 0.3 release (and in the sig-stacks sprint
retrospective recently), the topic came up of documenting the definition
of done, so that we can all be agreed about it. The difference between
the definition we had previously been using and the definition in this
changeset is that documentation and error reporting are now included.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>