docs: add error and event reporting one-pager
As we begin to fill out the troubleshooting niceties for users and
administrators of Crossplane, we will want to have a framework for
reporting errors and events. The document added by this changeset aims
to describe a mindset and framework that we can use to accomplish our
goals for errors and events.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>