Don't run workload on clusters without connection secrets
This was already a problem - we'd schedule workloads to a cluster that did not
specify a connection secret (and was thus unusable by Crossplane). Now that
connection secrets are pointer types this resulted in a nil pointer in the tests
which highlighted the fact that we should handle this condition.

Signed-off-by: Nic Cope <negz@rk0n.org>