design: add design doc for template stacks experience
As part of creating template stacks, we want to explore what the
experience looks like at the beginning, and then work backwards from
there. This helps us focus on the user experience early, and helps us
iron out **some** kinks before we even begin. This design doc and the
code which it links to are an attempt at that.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>