stacks: CRDs should use a managed-by label, not annotation
The `app.kubernetes.io/managed-by` annotations that Stack Manager adds to
Stack CRDs was based on a [Kubernetes recommendation for common
labels](https://kubernetes.io/docs/concepts/overview/working-with-objects/common-labels/).

The annotation should have been a label.

Additionally, the Stack design has been updated to reflect the use of
annotations and labels on CRDs.  The need for this documentation was
observed during the review of
https://github.com/crossplaneio/crossplane/pull/961 which will also
benefit from a more formal presentation of Stack Labels and Annotations.

Signed-off-by: Marques Johansson <marques@upbound.io>