template stacks experience: update to discuss stack.yaml and defaults
We've changed the way we want to specify the behaviors configuration and
the default values for configuration bindings. This updates the
experience design document to reflect it. It also updates the design
document to reference the speculative lifecycle hook design, and to
point to the internals design document in the repository now that it's
been merged.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>