stacks: #1049 clusterrole for namespaced stack deployments
This PR modifies StackInstall handling to create a ClusterRole rather
than a Role for use by the deployment controller.

Stacks (not just ClusterStacks) should use a ClusterRole for the
controller deployment ("system") roles in support of reuse for between
instances of a Stack in different namespaces.

The named Cluster Role should be reusable (well-known naming) and named
to permit instances of multiple versions of the same stack.

Closes #1049

Signed-off-by: Marques Johansson <marques@upbound.io>