template stacks experience: clarify, link to more examples
Now that we have a helm engine example and a quick start example, we
should link to them to help the reader find more information about
different scenarios. Also, some of the details of the examples in the
document are different now that I've been prototyping some of the
resource processing pipeline.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>