Aggregate Resources design document
This commit introduces a design document for aggregate resources and resource
classes. The design enables infrastructure operators to publish resource classes
that represent _groups_ of infrastructure, for example "an Azure SQL server with
three virtual network rules", or "a GKE cluster with two node pools in a new
subnet". The design maintains the flexible, composable, high-fidelity nature of
Crossplane managed resources and requires little to no change to existing APIs
and logic.

Signed-off-by: Nic Cope <negz@rk0n.org>