template stacks experience: assume stack.yaml is the only config
Some of the sample steps for writing a stack were referring to both the
app.yaml and the stack.yaml as though they were not yet merged; this has
been reworded to assume that they have already been merged by the time
the samnple steps are taken. A disclaimer has also been added to explain
this assumption.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>