template stacks experience: update examples
Some sections linked out to the experience repo quick start for
examples, and some sections linked out to the repo root. This updates
sections that linked out to the quick start to first show an example
inline and then link out to the quick start for more detail. It updates
sections that linked to the root of the repository to link to the quick
start section instead.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>