stacks: fix stack pull secrets and policy propagation to deployment
The Stack controller was creating deployments using an extra piece of
Stack spec defining imagePullSecrets and imagePullPolicy.

These extra pieces were not being set when the StackInstall handler created the
Stack resource, which created a bug that prevented the Deployment from
using the StackInstall specified imagePullSecrets and imagePullPolicy.

These fields can be considered extraneous because the StackInstall
embeds their values into the Stack (and StackDefinition) Controller
properties.

A user creeating a Stack resource by hand should not need to specify
these values outside of the Stack Controller fields where they are used.

The fix in this PR is to remove the extra fields from the Stack
resource and simply rely on the already prepared DeploymentSpec in the Stack
Controller.

Tests have been added to ensure that these Stack values are present in
the Deployment.

Signed-off-by: Marques Johansson <marques@upbound.io>