stacks: include apiVersion field in app.yaml
The apiVersion field of app.yaml expresses the version of the Stacks
spec that the package conforms to.  Changes to the spec that affect the
contents of the app.yaml file, including the replacement of the app.yaml
file, will be denoted in the Stack design spec (or its successors).

As stack packages adopt new standards they will embed the Stack
apiVersion in their metadata.

The change introduced here merely creates a new type,
PackageMetadataSpec, extending the AppMetadataSpec with an additional
APIVersion field.  The `app.yaml` file conforms to the
PackageMetadataSpec while the Stack.Spec confoms to the AppMetadataSpec.

The public representation of a Stack does not need to include the
internal packaging format from which that the Stack was derived.

No additional changes are included in this change as there are no
conflicting Stack versions to warrant content conversion or alternate
processing.

Signed-off-by: Marques Johansson <marques@upbound.io>