stack manager: inject stack controller image if not specified
We want to be able to inject the stack controller image if it isn't
specified, and we use the image from the stack envelope. This is because
a common pattern is to use the same image for both the envelope and the
controller, and if you need to specify the controller image with an
explicit version tag from inside the envelope image (which is itself),
it is inconvenient to manage the versions. It's much simpler to be able
to "inherit" the image and version from whatever was specified on the
stack install object.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>