stacks: truncate resource names and labels for host-aware environments
Resource names in the host environment may not exceed 253 characters.
Since these names are based on tenant namespace+name they could easily
exceed this limitation.

There are also some resources created on the host which invoke admission
controller behaviors adding resource labels that match resource names.
Because labels values are limitted to 63 characters the resource names
created on the host should not exceed 63 characters.

Because the names have been truncated, annotations have been added on
the host to help identify the related tenant resource when debugging.

This PR adds a truncation package with helpers for specific use-cases.
Truncation is performed in a predictable way to make it possible to find
resources by label knowing only the original resource names.  The suffix
of the truncation is a subset of a lowercased base32 checksum of a SHA1
digest of the original string.  SHA1's security limitations are not a
factor here because this is not being used for security purposes, it is
chosen for quick and even distribution in the checksum.

The suffix is lowercase because resource names must be lowercase.
Base32 was chosen to avoid "/" and a conflicting set of alphabetic
characters (upper and lower case are used in base64).

Signed-off-by: Marques Johansson <marques@upbound.io>

stacks: multi-parent labels should not exceed max name lengths

Signed-off-by: Marques Johansson <marques@upbound.io>