release doc: add template stacks
We maintain a set of template stacks, and we cut semver releases of our
stacks as part of the full release process for crossplane. So we want to
release the template stacks that we maintain also. They have a slightly
different process from existing stacks, so we need to add some guidance
to our documentation.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>