template-stacks: Remove Command field from Deployment that hosts the templating-controller
In the latest v0.3.0-rc version of the templating-controller, the entrypoint changed from
the typical kubebuilder /manager to a specific /templating-controller binary. To be
backwards compatible, the Stack Manager now just omits the Command (entrypoint) from the
deployment that it generates and expects the templating-controller image to have explicitly
defined it's entrypoint.

Signed-off-by: Jared Watts <jbw976@gmail.com>