Re-enable updating CRDs on new StackInstalls
Previously, the stack controller was updated to include the ability
to install multiple stacks that depend on the same CRD, and update
it to the newly installed version. This was broken when CRDs were
changed to be updated without directly modifying the existing CRD
(#1406). This re-enables the intended functionality, but is serving
as a stop-gap fix in a larger refactoring effort.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>