Check pod status multiple times in integration test
A successful integation test was observed in #1436 when
a failure was expected. This was reproduced locally and
observed that the test was not giving enough time for
the controller manager to start (and potentially fail)
before it reported successful. This update makes the
integration test check every 5 seconds for 25 seconds
to give the pods time to startup and stabilize.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>