Apply finalizers to stacks resources in main reconcile
Finalizers on stacks and stackinstalls were previously being
set at creation time. This meant that in a restore scenario
where a resource already existed on first reconciliation,
finalizers were never being set. Then, when the resource
was deleted, its children were not being cleaned up because
the parent already did not exist in the api server on the
subsequent reconciliation.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>