Add a string format transform
I added this transform to allow us to append/suffix values to existing strings,
but it could also be useful to convert other types (i.e. integers) to strings.

The below transform appends a suffix to the UID of the composite resource in
order to ensure the connection secret of the composed resource is written to
a unique secret name.

patches:
- fromFieldPath: "metadata.uid"
  toFieldPath: "spec.writeConnectionSecretToRef.name"
  transforms:
  - type: string
    string:
      fmt: "%s-sql"

Signed-off-by: Nic Cope <negz@rk0n.org>