Don't link to specific docs in examples
Linking to a specific version of the docs means we're linking to docs that will
become stale over time, and linking to master means we risk the specific page
we're linking to being removed or becoming irrelevant. Instead just link to the
docs and let the reader join the dots.

Signed-off-by: Nic Cope <negz@rk0n.org>