Use uppercase 'PostgreSQL' as Alibaba RDS engine version
I'm not sure whether the API is case sensitive, and I can't validate this due to
a lack of credentials, but I notice the API docs use uppercase engine versions.
We're using uppercase in the publishing infrastructure section of the getting
started guide, so if this is broken at least it will be consistently broken. :)

https://www.alibabacloud.com/help/doc-detail/26228.htm?spm=a2c63.p38356.879954.6.583c466afelHH0

Signed-off-by: Nic Cope <negz@rk0n.org>