Allow passing non-zero fsGroup in package deployments
Setting the fsGroup for a pod means that any volumes
that support ownership management are to be modified
to be owned and writable by the GID specified. In
addition, the GID will be added as a supplementary
group for containers in that pod. This is specifically
useful in the case where a projected Service Account
token is mounted, as Kubernetes currently sets the mode
to 0600 always, which may cause non-root containers to
be unable to read the file.

This change allows for packages to pass in a non-zero
fsGroup in their install.yaml. If an fsGroup of 0 is
passed it will be removed from the Pod Security Context."

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>