Merge XRP into XRD
This commit removes the CompositeResourcePublication (XRP) type, replacing it
with a field - spec.claimNames - on a CompositeResourceDefinition (XRD). An
XRD that specifies claimNames will define both a composite resource and a
composite resource claim - a namespaced CR that platform operators may use to
offer the composite resource to their platform consumers.

As a simplification I'm currently assuming that claim names are immutable once
set. They may be added retroactively (i.e. to offer a claim for a composite that
originally did not), but they may not be updated or removed once set. We could
change this in future at the expense of some additional complexity but without
breaking API changes if we so desired.

I've also kept the claim-CRD-and-controller-wrangling code in its own controller
in part because I prefer two quite complex reconcile loops to one really complex
reconcile loop, and in part because it was a smaller change to the current code
to do so.

Signed-off-by: Nic Cope <negz@rk0n.org>