Register pkg meta types with scheme
Crossplane does not register pkg meta types in its controller manager
scheme, but consumers of the crossplane-runtime package parser may want
to register the pkg meta types to the meta scheme they pass to the
parser. Currently, they are required to manually call Register for the
types to actually be registered with the scheme, which can be unexpected
because most other types are registered automatically in their init
function. Note that these types should never be registered with the
Crossplaen controllers because they are not actually installed into the
Kubernetes cluster.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>