Use camel case reasons for package events
Some of the package reconciler events were using error strings as reasons. API
conventions require event reasons to be camel case. In these cases I've wrapped
the event message (i.e. the upstream error) with the error text we were using as
a reason, which is our convention in most other reconcilers.

Signed-off-by: Nic Cope <negz@rk0n.org>