Remove the auto-target controller
This controller automatically generates KubernetesTargets for each
KuberntesCluster claim. We're marking KubernetesTarget deprecated in v0.13
along with KubernetesApplications (but not removing them). The KubernetesCluster
claim is being removed in v0.13, so this controller must go. Users will still be
able to manually create KubernetesTargets until they are removed in a future
release.

Signed-off-by: Nic Cope <negz@rk0n.org>