Add xpkg pkg for building and interacting with crossplane packages
Crossplane packages are opionated OCI image tarballs with the .xpkg
extension. This pkg contains utilities for building, naming, and parsing
both the on-disk crossplane package format and .xpkg tarballs directly.
If other crossplane pkgs are moved to an internal directory in the
future we should consider keeping this one public.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>