Replace package manager install pod manager with digester
Now that install pods are no longer used by package revisions there is
no need for the package controllers to create a pod only to fetch the
digest of the package image. The pod manager is now replaced with a
lighter weight digester which fetches an image digest used for
generating a package revision's name.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>