Add OCI image parser backend for parsing package contents
Replaces the pod log parser backend with an image backend that first
attempts to fetch an image from the xpkg cache, and if unsuccessful will
retrieve it from a registry. The xpkg cache is pluggable, meaning that
you can modify the image backend behavior by providing a cache with
different properties. This is most clearly illustrated by providing a
NopCache, which will cause the image backend to re-fetch the package
image on every initialization.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>