Allow for package revision controllers to use pre-cached images
In the event that a package is installed with pullPolicy Never we will
only try to get it from the cache and will not attempt to fetch from a
registry. The exact name of the spec.package field will be used to
identify the pre-cached image. This differs from the typical case where
we fetch the image and write the xpkg tarball with a name that matches
the package revision. A side-effect of users being able to supply their
own pre-cached images is that we will not remove them from the cache if
the revision that uses them is deleted, unless the cached xpkg name
matches the name of the package revision, which is
<provider-name>-<xpkg-name> when the pullPolicy is Never.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>