Default to current for crank install version
Updates the install.sh script to default to current rather than a pinned
stable version of crossplane. This means that if CHANNEL is master and
no VERSION is supplied then we will get latest master build. If CHANNEL
is alpha and no VERSION is supplied then we will get latest stable.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>