Make getting started with composition less intimidating
...by using our new configuration packages to install the XRDs and Compositions!
I'm hoping to follow up with a section on how to create these configuration
packages if there's time before the release - I've created a snippets/package
directory so that we don't lose the examples that are used to build the config
packages. Currently the packages are just built and pushed with kubectl
crossplane, though it would be great if they were automatically built by CI in
future.

Signed-off-by: Nic Cope <negz@rk0n.org>