Close cached package files after write
File descriptors can be leaked when a file is opened and not subsequently
closed. In the case of the package manager, there would likely need to be a
large number of packages installed before this became an issue as there
should only be one open per revision when it is initially stored. Files
are opened and closed correctly currently on Get calls.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>