Aggregate edit and view roles to admin
This updates the crossplane-admin role to be given any roles that are
given to the edit or view roles. We still explicitly aggregate to the
admin when the RBAC manager creates roles, so these aggregation lables
are convenience for if a user want to manually create a role for edit or
view that should roll up to admin.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>