Port CI/CD to GitHub Actions
This commit is a moderately faithful port of the existing Jenkinsfile to GitHub
Actions. Notably, it eschews the cross container to avoid the ~5 min penalty of
having to build it and runs much of the build in parallel.

Signed-off-by: Nic Cope <negz@rk0n.org>
(cherry picked from commit be07dd765e5c1044ef704967a68ef7d879759f41)