Treat package as sole owner of its revisions
Updates the package manager behavior to dictate the state on its
revision on every reconile. This allows for more accurate status
messages, such as not emitting an unhealthy status when the revision has
just not completed install. Also allows for changes of the package spec
to flow down to the revision. This can be useful in the case where you
want to fix a currently failing active revision by adjusting something
like its pull secrets or ignoring its constraints.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>