Make package friendly indentifier a valid DNS label name
Updates FriendlyID to replace all periods in returned identifier such
that a valid DNS subdomain name is converted to a valid DNS label. This
is important for the case in which the identifier is used as the name
for a automounted service account volume name, which must adhere to DNS
label standards.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>