Update package manager revisioner to make head request
The package revisioner only needs to acquire the the image digest to
generate a new revision name. Currently, this is done by making a GET
request, but the same functionality could be achieved with a HEAD
request, which is lighter weight and does not count as an image pull.
This updates the revisioner to use make a HEAD request.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>