Persist references to composed resources before creating them
This commit refactors the XR reconciler such that new composed resources are
named (by the API server via a dry-run create) before they are created. This
allows us to persist a reference the new composed resource before it is created,
which in turn ensures we don't leak composed resources if we fail to update our
references. Instead we'll notice that we already named the composed resource in
question (by reading our references) and use that name.

Fixes https://github.com/crossplane/crossplane/issues/1910

Signed-off-by: Nic Cope <negz@rk0n.org>