Watch for service accounts and CRDs owned by provider revisions
I recently encountered a ClusterRoleBinding for a ProviderRevision that had no
subject. I suspect this is because there was enough of a delay between when the
ProviderRevision was created and when its ServiceAccount was created that this
controller listed ServiceAccounts before the ProviderRevision's ServiceAccount
was created. This commit should cause the ProviderRevision to be queued for a
reconcile when its ServiceAccount is added, or when more CRDs are added.

Signed-off-by: Nic Cope <negz@rk0n.org>