Allow XRs and XRCs to record when they last published connection details
In addition to being useful debug information, this allows a controller to
watch for changes to connection details indirectly by watching the resource
those connection details pertain to; i.e. the XRC controller will have a
reconcile queued when the XR's connection details change. It's possible today
for the XRC controller to instead watch the XR's connection secret directly, but
this functionality could be useful if and when connection details are published
to an external system (e.g. Vault).

Signed-off-by: Nic Cope <negz@rk0n.org>