Migrate the ccrds package out of apiextensions/v1alpha1
The ccrd package is used to generate the underlying CRDs for XRDs. It's coupled
to a specific version of the XRD types (e.g. v1alpha1) but in practice we would
not replicate it if we supported many XRD types - we'd only implement it for the
version that our controllers reconciled.

Signed-off-by: Nic Cope <negz@rk0n.org>