Add v1beta1 iterations of all core APIs
These APIs are identical to the existing v1alpha1 APIs, and can thus be served
concurrently without conversion webhooks. Note that the upstream backward
compatibility gotchas section mentions that we should not do what we're doing in
this commit; introduce a new version and make it the storage version. I assume
this rule only applies when the new version includes breaking changes that might
need to be would back, and thus it is safe to bump the storage version in our
case. Bumping the storage version now increases the likelihood that users will
automatically have their storage versions upgraded to v1beta1 in time for the
next Crossplane release, which should allow us to deprecate the v1alpha1 APIs.

https://github.com/kubernetes/community/blob/46f114/contributors/devel/sig-architecture/api_changes.md#backward-compatibility-gotchas

Signed-off-by: Nic Cope <negz@rk0n.org>