Do not watch CRDs in RBAC manager provider revision controller
The ProviderRevision controller updates the CRDs every minute, which can
enqueue many requests for the RBAC manager provider revision controller.
In practice, the addition or removal of a CRD for a package will always
be reflected in the ProviderRevision, so the RBAC manager will have a
reconcile triggered within 60 seconds if any changes are made.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>