Do not append hash when xpkg name provided
Currently, if a custom name is used when building an xpkg rather than
assuming the name in the crossplane.yaml, the digest of the package is
appended to the custom name. We should assume that a user wants to use
the custom name exactly as they have provided and not append a hash in
that case. This updates to make that change.

When the name in crossplane.yaml is used, the image hash will still be
appended.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>