Add dependency resolver controller
Adds a new controller for resolving dependencies of packages. It watches
for revisions and the package lock and creates new packages when a
package revision adds itself to the lock with unsatisfied dependencies.

It also adds a finalizer on the package lock. This provides an extra
guard against a user accidentally deleting the package lock. However,
this is not the primary intended use case of a finalizer, and it could
cause unexpected behavior due to the fact that it is never removed by
the controller. In the future, Crossplane may instead create the lock if
it is missing in the cluster.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>