Add named Patch Sets to Compositions
Implements a new patchSets field on CompositionSpec for users to define
groupings of Patches at the composite level against a name.

We also introduce the idea of each patch having a type field (defaulted
to FromCompositeFieldPath).

When a patch with type PatchSet is defined on a composed resource,
patches from a matching PatchSet defined on the composite resource will
be inlined in its' place.

Signed-off-by: Ben Agricola <bagricola@squiz.co.uk>