Ensure valid DNS label for FriendlyID
Updates FriendlyID package revision naming function to always return a
valid DNS label made up of a concatenation of the package name and the
revision digest. This will not result in changes for the typical naming
scenario, but can guard against issues when a revision image has been
manually loaded into the cache with an unexpected name.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>