Do not default registry name when resolving dependencies
This updates the logic that adds package revisions to the Lock to avoid
defaulting to a registry. Registry defaulting still takes place when
actually pulling an image, but it remains opaque to users. All
dependencies and revisions will be added to the Lock in the same format
they were declared.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>