Convert meta.pkg.crossplane.io types to v1 fewer times
The reconciler passes the entire package contents to the linter before we
extract the metadata, so we need to convert it in a few places. Adding a
conversion here will ensure _fewer_ conversions though - e.g. the Crossplane
compatibility check, dependency resolution, and hooks won't convert the meta
types a second time (they'll still try just in case, but will short circuit
because the converted type is no longer a conversion.Convertible).

Signed-off-by: Nic Cope <negz@rk0n.org>