Install latest semantic version of providers
We currently install providers based on the alpha release channel to
avoid churn on docs at release time. As we move towards more
independence in terms of maintainership for providers, we should allow
provider maintainers to update documentation to install with the latest
semantic version of their provider, which we encourage as best practice
when dealing with packages.

This also allows us to have a simpler getting started workflow with
first installing providers, then later installing configuration packages
that depend on them. This workflow is currently broken because the
configuration packages do not parse alpha as a valid semantic version.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>