Enable more kinds of updates to Compositions
Currently we use 'anonymous' resource templates. These templates rely on
matching the order of the composition resource template array relative to the
composite resource references array in order to match composed resources to the
templates used to render them. This means we can't reorder the templates within
a composition (or change to using a Composition with a different template order)
without affecting composed resources.

This commit introduces optional 'named' templates. If all templates within a
Composition are named Crossplane will use the names to relate templates to
composed resources. This allows us to reorder templates, and to delete composed
resources when their corresponding template is deleted.

Signed-off-by: Nic Cope <negz@rk0n.org>